feat: optimize blog cover (using static process images)
