--- public/fonts/arial-bold.otf
@@ -1,1 +1,0 @@
-
