--- public/fonts/arial.otf
@@ -1,1 +1,0 @@
-
